split correspondace service into "receipt" and "call to action" services
