+++ src/Altinn.App.Core/Extensions/ServiceCollectionExtensions.cs
@@ -305,7 +305,8 @@ public static class ServiceCollectionExtensions
     {
         services.AddHttpClient<ISignClient, SignClient>();
         services.AddTransient<ISigningDelegationService, SigningDelegationService>();
-        services.AddTransient<ISigningCorrespondenceService, SigningCorrespondenceService>();
+        services.AddTransient<ISigningReceiptService, SigningReceiptService>();
+        services.AddTransient<ISigningCallToActionService, SigningCallToActionService>();
         services.AddTransient<ISigningService, SigningService>();
     }
