+++ src/Altinn.App.Core/Features/Signing/Models/SigneeParty.cs
@@ -94,8 +94,8 @@ public class Sms
     /// <summary>
     /// The body. If not set, a default will be used.
     /// </summary>
-    [JsonPropertyName("body")]
-    public string? Body { get; set; }
+    [JsonPropertyName("textResourceKey")]
+    public string? TextResourceKey { get; set; }
 
     /// <summary>
     /// The reference used to track the sms. Can be set to a custom value. If not set, a random guid will be used.
@@ -117,14 +117,14 @@ public class Email
     /// <summary>
     /// The subject. If not set, a default will be used.
     /// </summary>
-    [JsonPropertyName("subject")]
-    public string? Subject { get; set; }
+    [JsonPropertyName("subjectTextResourceKey")]
+    public string? SubjectTextResourceKey { get; set; }
 
     /// <summary>
     /// The body. If not set, a default will be used.
     /// </summary>
-    [JsonPropertyName("body")]
-    public string? Body { get; set; }
+    [JsonPropertyName("bodyTextResourceKey")]
+    public string? BodyTextResourceKey { get; set; }
 
     /// <summary>
     /// The reference used to track the email. Can be set to a custom value. If not set, a random guid will be used.
